--- a/frontends/infinite-corridor/src/foo.test.ts
+++ b/frontends/infinite-corridor/src/foo.test.ts
@@ -1,0 +1,3 @@
+test("Addition", () => {
+  expect(1 + 1).toBe(2)
+})