--- a/frontends/infinite-corridor/src/pages/field-details/EditFieldAppearanceForm.tsx
+++ b/frontends/infinite-corridor/src/pages/field-details/EditFieldAppearanceForm.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { useHistory } from "react-router"
 import { Formik, Form, Field } from "formik"
-import { Button } from "@mui/material"
+import Button from "@mui/material/Button"
 import * as Yup from "yup"
 
 import { FieldChannelAppearanceForm, useMutateField } from "../../api/fields"