--- a/frontends/infinite-corridor/src/pages/resource-lists/AddToListDialog.test.tsx
+++ b/frontends/infinite-corridor/src/pages/resource-lists/AddToListDialog.test.tsx
@@ -1,4 +1,4 @@
-import { faker } from "@faker-js/faker"
+import { faker } from "@faker-js/faker/locale/en"
 import {
   makeCourse,
   makeUserListsPaginated,