--- a/frontends/infinite-corridor/src/pages/resource-lists/ManageListDialogs.test.tsx
+++ b/frontends/infinite-corridor/src/pages/resource-lists/ManageListDialogs.test.tsx
@@ -1,4 +1,4 @@
-import { faker } from "@faker-js/faker"
+import { faker } from "@faker-js/faker/locale/en"
 import { pick } from "lodash"
 import {
   LearningResourceType as LRT,