--- a/frontends/infinite-corridor/src/pages/resource-lists/ResourceListsListingsPage.test.tsx
+++ b/frontends/infinite-corridor/src/pages/resource-lists/ResourceListsListingsPage.test.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { faker } from "@faker-js/faker"
+import { faker } from "@faker-js/faker/locale/en"
 import {
   Favorites,
   LearningResourceCardTemplate as LRCardTemplate,