--- a/frontends/ol-util/src/components/TitledCarousel.test.tsx
+++ b/frontends/ol-util/src/components/TitledCarousel.test.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable testing-library/no-node-access */
 import React from "react"
-import { faker } from "@faker-js/faker"
+import { faker } from "@faker-js/faker/locale/en"
 import { render, screen } from "@testing-library/react"
 import user from "@testing-library/user-event"
 import TitledCarousel, { TitledCarouselProps } from "./TitledCarousel"