--- a/frontends/ol-util/src/components/embedly/EmbedlyCard.test.tsx
+++ b/frontends/ol-util/src/components/embedly/EmbedlyCard.test.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable testing-library/no-node-access */
 import React from "react"
-import { faker } from "@faker-js/faker"
+import { faker } from "@faker-js/faker/locale/en"
 import { assertInstanceOf } from "ol-util"
 import { render, screen, waitFor } from "@testing-library/react"
 import EmbedlyCard, { EmbedlyCardProps } from "./EmbedlyCard"