--- a/jest-shared-setup.ts
+++ b/jest-shared-setup.ts
@@ -2,15 +2,7 @@ import failOnConsole from "jest-fail-on-console"
 import '@testing-library/jest-dom'
 import { configure } from "@testing-library/dom"
 
-failOnConsole({
-  /**
-   * This warning is removed from newer versions of React because it is
-   * misleading (not usually a memory leak).
-   *
-   * See https://github.com/reactwg/react-18/discussions/82
-   */
-  silenceMessage: msg => /Can't perform a React state update on an unmounted component/.test(msg),
-})
+failOnConsole()
 
 configure({
   /**